Cria autorização baseada nas permissões

+++ app/Http/Controllers/ClienteController.php
@@ -8,29 +8,27 @@ use Illuminate\Http\Request;
 
 class ClienteController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     */
     public function index()
     {
+        if (!auth()->user()->tokenCan('cliente-index')) {
+            return Response()->json([
+                "message" => "Sem permissão"
+            ], 403);
+        }
+
         $dados = Cliente::get();
 
         return Response()->json($dados);
     }
 
-    /**
-     * Show the form for creating a new resource.
-     */
-    public function create()
-    {
-        //
-    }
-
-    /**
-     * Store a newly created resource in storage.
-     */
     public function store(Request $request)
     {
+        if (!auth()->user()->tokenCan('cliente-store')) {
+            return Response()->json([
+                "message" => "Sem permissão"
+            ], 403);
+        }
+
         $dados = $request->except('_token');
 
         $cliente = Cliente::create($dados);
@@ -38,11 +36,14 @@ class ClienteController extends Controller
         return Response()->json($cliente, 201);
     }
 
-    /**
-     * Display the specified resource.
-     */
     public function show(string $id)
     {
+        if (!auth()->user()->tokenCan('cliente-show')) {
+            return Response()->json([
+                "message" => "Sem permissão"
+            ], 403);
+        }
+
         $cliente = Cliente::find($id);
 
         if(!$cliente) {
@@ -52,19 +53,14 @@ class ClienteController extends Controller
         return Response()->json($cliente);
     }
 
-    /**
-     * Show the form for editing the specified resource.
-     */
-    public function edit(string $id)
-    {
-        //
-    }
-
-    /**
-     * Update the specified resource in storage.
-     */
     public function update(Request $request, string $id)
     {
+        if (!auth()->user()->tokenCan('cliente-update')) {
+            return Response()->json([
+                "message" => "Sem permissão"
+            ], 403);
+        }
+
         $dados = $request->except('_token');
 
         $cliente = Cliente::find($id);
@@ -74,11 +70,14 @@ class ClienteController extends Controller
         return Response()->json([], 204);
     }
 
-    /**
-     * Remove the specified resource from storage.
-     */
     public function destroy(string $id)
     {
+        if (!auth()->user()->tokenCan('cliente-destroy')) {
+            return Response()->json([
+                "message" => "Sem permissão"
+            ], 403);
+        }
+
         Cliente::destroy($id);
 
         return Response()->json([], 204);
