+++ app/Http/Controllers/VendaController.php
@@ -8,29 +8,27 @@ use Illuminate\Http\Request;
 
 class VendaController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     */
     public function index()
     {
+        if (!auth()->user()->tokenCan('venda-index')) {
+            return Response()->json([
+                "message" => "Sem permissão"
+            ], 403);
+        }
+
         $dados = Venda::get();
 
         return Response()->json($dados);
     }
 
-    /**
-     * Show the form for creating a new resource.
-     */
-    public function create()
-    {
-        //
-    }
-
-    /**
-     * Store a newly created resource in storage.
-     */
     public function store(Request $request)
     {
+        if (!auth()->user()->tokenCan('venda-store')) {
+            return Response()->json([
+                "message" => "Sem permissão"
+            ], 403);
+        }
+
         $dados = $request->except('_token');
 
         $venda = Venda::create($dados);
@@ -38,19 +36,14 @@ class VendaController extends Controller
         return Response()->json($venda, 201);
     }
 
-    /**
-     * Display the specified resource.
-     */
     public function show(string $id)
     {
-        //
-    }
+        if (!auth()->user()->tokenCan('venda-show')) {
+            return Response()->json([
+                "message" => "Sem permissão"
+            ], 403);
+        }
 
-    /**
-     * Show the form for editing the specified resource.
-     */
-    public function edit(string $id)
-    {
         $venda = Venda::find($id);
 
         if(!$venda) {
@@ -60,11 +53,14 @@ class VendaController extends Controller
         return Response()->json($venda);
     }
 
-    /**
-     * Update the specified resource in storage.
-     */
     public function update(Request $request, string $id)
     {
+        if (!auth()->user()->tokenCan('venda-update')) {
+            return Response()->json([
+                "message" => "Sem permissão"
+            ], 403);
+        }
+
         $dados = $request->except('_token');
 
         $venda = Venda::find($id);
@@ -74,11 +70,14 @@ class VendaController extends Controller
         return Response()->json([], 204);
     }
 
-    /**
-     * Remove the specified resource from storage.
-     */
     public function destroy(string $id)
     {
+        if (!auth()->user()->tokenCan('venda-destroy')) {
+            return Response()->json([
+                "message" => "Sem permissão"
+            ], 403);
+        }
+
         Venda::destroy($id);
 
         return Response()->json([], 204);
